// frontend/ticket-booking-web/src/app/api/model/paged-result-of-ticket-dto.ts
/**
 * Ticket Booking Platform API - v1
 *
 * Contact: m1musbal@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TicketDto } from './ticket-dto';


export interface PagedResultOfTicketDto { 
    items: Array<TicketDto>;
    page: number;
    pageSize: number;
    totalCount: number;
    totalPages?: number;
}