// frontend/ticket-booking-web/src/app/api/model/ticket-status.ts
/**
 * Ticket Booking Platform API - v1
 *
 * Contact: m1musbal@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export const TicketStatus = {
    Valid: 'Valid',
    Used: 'Used',
    Cancelled: 'Cancelled'
} as const;
export type TicketStatus = typeof TicketStatus[keyof typeof TicketStatus];